Use pgoutput instead of triggers for mtrees
* Add mtree init, teardown, and listen
* Use a separate file for SQL templates
* Update auth to get a replication connection
