+++ ace.yaml
@@ -23,10 +23,15 @@ table_diff:
   max_diff_batch_size: 1000
   compare_unit_size: 10000
 
-mtree_diff:
-  min_block_size: 1000
-  block_size: 100000
-  max_block_size: 1000000
+mtree:
+  cdc:
+    slot_name: "ace_mtree_slot"
+    publication_name: "ace_mtree_pub"
+
+  diff:
+    min_block_size: 1000
+    block_size: 100000
+    max_block_size: 1000000
 
 server:
   listen_address: "0.0.0.0"
