+++ internal/auth/auth.go
@@ -16,28 +16,54 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/jackc/pgx/v5/pgconn"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-func GetClusterNodeConnection(nodeInfo map[string]any, clientRole string) (*pgxpool.Pool, error) {
+func getConnStr(nodeInfo map[string]any) string {
 	user, _ := nodeInfo["DBUser"].(string)
 	password, _ := nodeInfo["DBPassword"].(string)
-	host, _ := nodeInfo["PublicIP"].(string)
+	var host string
+	if h, ok := nodeInfo["Host"].(string); ok && h != "" {
+		host = h
+	} else if h, ok := nodeInfo["PublicIP"].(string); ok && h != "" {
+		host = h
+	}
 	database, _ := nodeInfo["DBName"].(string)
-	port, okPort := nodeInfo["Port"].(string)
 
-	if !okPort {
+	var port string
+	if p, ok := nodeInfo["Port"].(string); ok {
+		port = p
+	} else if p, ok := nodeInfo["Port"].(float64); ok {
+		port = fmt.Sprintf("%.0f", p)
+	}
+
+	if port == "" {
 		port = "5432"
 	}
 
-	connStr := fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=disable", user, password, host, port, database)
+	return fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=disable", user, password, host, port, database)
+}
+func GetClusterNodeConnection(nodeInfo map[string]any, clientRole string) (*pgxpool.Pool, error) {
+	connStr := getConnStr(nodeInfo)
 
-	pool, err := SetupDBPool(context.Background(), connStr, fmt.Sprintf("%s:%s", host, port))
+	pool, err := SetupDBPool(context.Background(), connStr, nodeInfo["Name"].(string))
 	if err != nil {
 		return nil, err
 	}
 
 	return pool, nil
+}
+
+func GetReplModeConnection(nodeInfo map[string]any) (*pgconn.PgConn, error) {
+	connStr := getConnStr(nodeInfo)
+	connStr = connStr + "&replication=database"
+
+	conn, err := pgconn.Connect(context.Background(), connStr)
+	if err != nil {
+		return nil, err
+	}
+	return conn, nil
 }
 
 func SetupDBPool(ctx context.Context, connStr, name string) (*pgxpool.Pool, error) {
